pyproject.toml
[project]
name = "schorlarly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "arxiv>=2.4.0",
    "dacite>=1.9.2",
    "faiss-cpu>=1.13.2",
    "fastapi[all]>=0.128.0",
    "google-genai>=1.56.0",
    "loguru>=0.7.3",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "perplexityai>=0.22.2",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "requests>=2.32.5",
    "ruff>=0.14.10",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.8",
]
